run test server serperate port cypress port cy:dev egen port reset

+++ prisma/reset.ts
@@ -1,6 +1,8 @@
 import { db } from "./db";
 
 async function main() {
+// ALDRIG LÅTA DETTA FUNKA UTANFÖR EN TEST MIJLÖ FÖR ATT INTE RÅKA TA BORT DATABASEN
+  if (process.env.NODE_ENV !== "test") return; 
   await db.post.deleteMany({});
 }
 main()
